refactor: move tableTypes to `src/components/Table/types.d.ts`
